refactor: RacingGame의 TryCount를 VO로 변경
