+++ src/main/java/racingcar/domain/racing/RacingGame.java
@@ -1,26 +1,27 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 import java.util.List;
-import racingcar.domain.car.Car;
+import racingcar.domain.IndexModel;
+import racingcar.domain.racing.car.Car;
 
 public class RacingGame extends IndexModel {
 
     private Participations participations;
-    private int tryCount;
+    private TryCount tryCount;
     private Winners winners;
 
-    private RacingGame(Participations participations, int tryCount, Winners winners) {
+    private RacingGame(Participations participations, TryCount tryCount, Winners winners) {
         this.participations = participations;
         this.tryCount = tryCount;
         this.winners = winners;
     }
 
-    public static RacingGame create(Participations participations, int tryCount, Winners winners) {
+    public static RacingGame create(Participations participations, TryCount tryCount, Winners winners) {
         return new RacingGame(participations, tryCount, winners);
     }
 
     public int getTryCount() {
-        return tryCount;
+        return tryCount.getTryCount();
     }
 
     public List<Car> getWinnerList() {
