+++ src/main/java/racingcar/domain/racing/Winners.java
@@ -1,8 +1,8 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 import java.util.ArrayList;
 import java.util.List;
-import racingcar.domain.car.Car;
+import racingcar.domain.racing.car.Car;
 
 public class Winners { // 일급컬렉션
