+++ src/main/java/racingcar/repository/MemoryCarRepository.java
@@ -1,6 +1,6 @@
 package racingcar.repository;
 
-import racingcar.domain.car.Car;
+import racingcar.domain.racing.car.Car;
 
 public class MemoryCarRepository extends MemoryRepository<Car> implements CarRepository {
