+++ src/main/java/racingcar/repository/MemoryRacingGameRepository.java
@@ -1,6 +1,6 @@
 package racingcar.repository;
 
-import racingcar.domain.RacingGame;
+import racingcar.domain.racing.RacingGame;
 
 public class MemoryRacingGameRepository extends MemoryRepository<RacingGame> implements RacingGameRepository {
