+++ src/test/java/racingcar/domain/ParticipationTest.java
@@ -4,7 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.Car;
+import racingcar.domain.racing.car.Car;
+import racingcar.domain.racing.Participations;
 
 public class ParticipationTest {
